# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: "0.0.1"
  title: Demo swagger app - get distance between 2 geo points
# during dev, should point to your local machine
host: 127.0.0.1:10010
# basePath prefixes all resource paths 
basePath: /api
# 
schemes:
  # tip: remove http to make production-grade
  - http
#  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /list:
    # binds a127 app logic to a route
    x-swagger-router-controller: test_api
    get:
      description: Returns all incoming requests in descending order
      # used as the method name of the controller
      operationId: list
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PointsListResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /add:
     x-swagger-router-controller: test_api
     post:
      description: Receives array of 2 points and calculates distance between them
      operationId: addPoint
      parameters:
        - name: points
          in: body
          description: JSON object with array of 2 geo points
          required: true
          schema: 
            $ref: "#/definitions/PointsRequest"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AddPointsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
          
# complex objects have schema definitions
definitions:
  Point:
    type: object
    required:
      - lat
      - lon
    properties:
      lat:
        type: number
        format: double
      lon:
        type: number
        format: double
  PointsRequest:
    type: object
    required:
        - points
    properties:
      points:
        type: array
        items: 
          type: object
          $ref: "#/definitions/Point"
    example:
      application/json:
        {
          "points": [{"lat": 44.960278, "lon": 34.114679}, {"lat": 44.962123, "lon": 34.120322}]
        }
  AddPointsResponse:
    required:
      - distance
      - measure
    properties:
      distance:
        type: number
        format: double
      measure:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  PointsListResponse:
    required:
      - points
    properties:
      points:
        type: array
        items:
          type: object